se mueve para adelante, gira y dispara

+++ app/src/main/java/starships/collidable/Asteroid.java
@@ -6,7 +6,7 @@ public class Asteroid extends Colisionable {
     private int initialHealth;
     private int currentHealth;
 
-    public Asteroid(String id, Position position, int rotationInDegrees, int height, int width, int direction, boolean clockwise, int initialHealth, int currentHealth) {
+    public Asteroid(String id, Position position, double rotationInDegrees, double height, double width, double direction, boolean clockwise, int initialHealth, int currentHealth) {
         super(id, CollidableType.ASTEROID, position, rotationInDegrees, height, width, CollidableShape.ELLIPTICAL, direction);
         this.clockwise = clockwise;
         this.initialHealth = initialHealth;
@@ -34,7 +34,7 @@ public class Asteroid extends Colisionable {
     private Colisionable move() {
         int newX = (int) (getPosition().getX() + 0.7 * Math.sin(Math.PI * 2 * getDirection() / 360));
         int newY = (int) (getPosition().getY() + 0.7 * Math.cos(Math.PI * 2 * getDirection() / 360));
-        int newRotationInDegrees;
+        double newRotationInDegrees;
         Position newPosition = new Position(newX, newY);
         if (clockwise) {
             newRotationInDegrees = getRotationInDegrees() + 2;
@@ -48,7 +48,7 @@ public class Asteroid extends Colisionable {
         this.clockwise = clockwise;
     }
 
-    public int calculateHealth(){
+    public double calculateHealth(){
         return (this.getWidth() * this.getHeight())/100;
     }
